+++ src/config/structs.rs
@@ -8,19 +8,15 @@ pub struct Config {
     pub active_profile: usize,
     /// The index of the active modpack
     pub active_modpack: usize,
-    /// The profiles
     pub profiles: Vec<Profile>,
-    /// The modpacks
     pub modpacks: Vec<Modpack>,
 }
 
 #[derive(Deserialize, Serialize, Debug, Clone)]
 pub struct Modpack {
-    /// The modpack's name
     pub name: String,
     /// The Minecraft instance directory to install to
     pub output_dir: PathBuf,
-    /// Whether to install overrides
     pub install_overrides: bool,
     /// The project ID of the modpack
     pub identifier: ModpackIdentifier,
@@ -34,23 +30,20 @@ pub enum ModpackIdentifier {
 
 #[derive(Deserialize, Serialize, Debug, Clone)]
 pub struct Profile {
-    /// The profile's name
     pub name: String,
     /// The directory to download mod files to
     pub output_dir: PathBuf,
-    /// Only download if the mod file is compatible with this Minecraft version
+    /// Only download mod files compatible with this Minecraft version
     pub game_version: String,
-    /// Only download  if the mod file is compatible with this mod loader
+    /// Only download mod files compatible with this mod loader
     pub mod_loader: ModLoader,
-    /// A list of all the mods configured
     pub mods: Vec<Mod>,
 }
 
 #[derive(Deserialize, Serialize, Debug, Clone)]
 pub struct Mod {
-    /// The name the mod
     pub name: String,
-    /// Identify the mod based on a mod source
+    /// The project ID of the mod
     pub identifier: ModIdentifier,
     /// Whether to check for game version compatibility
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -76,9 +69,9 @@ pub enum ModLoader {
 #[derive(thiserror::Error, Debug, PartialEq, Eq)]
 #[error("The given string is not a mod loader")]
 pub struct ModLoaderParseError {}
-impl TryFrom<&String> for ModLoader {
+impl TryFrom<&str> for ModLoader {
     type Error = ModLoaderParseError;
-    fn try_from(from: &String) -> Result<Self, Self::Error> {
+    fn try_from(from: &str) -> Result<Self, Self::Error> {
         match from.to_lowercase().as_str() {
             "quilt" => Ok(Self::Quilt),
             "fabric" => Ok(Self::Fabric),