+++ src/file_picker.rs
@@ -1,37 +1,53 @@
-use std::path::{Path, PathBuf};
+use crate::HOME;
+use std::{
+    env::current_dir,
+    io::Result,
+    path::{Path, PathBuf},
+};
 
-// macOS, it can only use a synchronous file picker, and any GUI feature
-#[cfg(all(target_os = "macos", any(feature = "gtk", feature = "xdg")))]
-#[allow(clippy::unused_async)]
-/// Use the file picker to pick a file, with a `default` path ([XDG not supported](https://github.com/PolyMeilex/rfd/issues/42))
-pub async fn pick_folder(default: &Path, prompt: &str) -> Option<PathBuf> {
+#[cfg(feature = "gui")]
+/// Use the system file picker to pick a file, with a `default` path that is [not supported on XDG](https://github.com/PolyMeilex/rfd/issues/42)
+fn show_file_picker(default: &Path, prompt: &str) -> Option<PathBuf> {
     rfd::FileDialog::new()
         .set_directory(default)
         .set_title(prompt)
         .pick_folder()
 }
 
-// Not macOS, other OSs can use the async version, and any GUI feature
-#[cfg(all(not(target_os = "macos"), any(feature = "gtk", feature = "xdg")))]
-/// Use the file picker to pick a file, with a `default` path ([XDG not supported](https://github.com/PolyMeilex/rfd/issues/42))
-pub async fn pick_folder(default: &Path, prompt: &str) -> Option<PathBuf> {
-    rfd::AsyncFileDialog::new()
-        .set_directory(default)
-        .set_title(prompt)
-        .pick_folder()
-        .await
-        .map(|handle| handle.path().into())
-}
-
-// No GUI features
-#[cfg(not(any(feature = "gtk", feature = "xdg")))]
-#[allow(clippy::unused_async)]
-pub async fn pick_folder(default: &Path, prompt: &str) -> Option<PathBuf> {
+#[cfg(not(feature = "gui"))]
+/// Use a terminal input to pick a file, with a `default` path
+fn show_file_picker(default: &Path, prompt: &str) -> Option<PathBuf> {
     use dialoguer::{theme::ColorfulTheme, Input};
     Input::with_theme(&ColorfulTheme::default())
-        .with_prompt(prompt)
         .default(default.display().to_string())
+        .with_prompt(prompt)
+        .report(false)
         .interact()
         .ok()
         .map(Into::into)
+}
+
+pub fn pick_folder(default: &Path, prompt: &str) -> Result<Option<PathBuf>> {
+    let input = show_file_picker(default, prompt);
+    Ok(match input {
+        Some(input) => {
+            let mut path = PathBuf::new();
+            let components = input.components();
+            for c in components {
+                path = path.join(if c.as_os_str() == "~" {
+                    HOME.to_owned()
+                } else if c.as_os_str() == "." {
+                    current_dir()?
+                } else {
+                    PathBuf::from(c.as_os_str())
+                });
+            }
+            println!(
+                "✔ \x1b[01mOutput Directory\x1b[0m · \x1b[32m{}\x1b[0m",
+                path.display()
+            );
+            Some(path)
+        },
+        None => None,
+    })
 }