+++ src/modpack/add.rs
@@ -5,7 +5,6 @@ use ferinth::{
 };
 use furse::{structures::mod_structs::Mod, Furse};
 use reqwest::StatusCode;
-use std::{sync::Arc};
 
 type Result<T> = std::result::Result<T, Error>;
 #[derive(thiserror::Error, Debug)]
@@ -53,46 +52,26 @@ impl From<ferinth::Error> for Error {
 /// Check if the project of `project_id` exists and is a modpack
 ///
 /// Returns the project struct
-pub async fn curseforge(
-    curseforge: Arc<Furse>,
-    config: &mut Config,
-    project_id: i32,
-) -> Result<Mod> {
+pub async fn curseforge(curseforge: &Furse, config: &Config, project_id: i32) -> Result<Mod> {
     let project = curseforge.get_mod(project_id).await?;
     // Check if project has already been added
     if config.modpacks.iter().any(|modpack| {
         modpack.name == project.name
             || ModpackIdentifier::CurseForgeModpack(project.id) == modpack.identifier
     }) {
-        return Err(Error::AlreadyAdded);
-    }
-
-    let files = curseforge.get_mod_files(project.id).await?;
-    let mut contains_zip_file = false;
-
-    // Check if the files are zip files
-    for file in &files {
-        if file.file_name.contains("zip") {
-            contains_zip_file = true;
-            break;
-        }
-    }
-
-    if contains_zip_file {
-        Ok(project)
-    } else {
+        Err(Error::AlreadyAdded)
+    // Check if the project is a modpack
+    } else if !project.links.website_url.as_str().contains("modpack") {
         Err(Error::NotAModpack)
+    } else {
+        Ok(project)
     }
 }
 
 /// Check if the project of `project_id` exists and is a modpack
 ///
 /// Returns the project struct
-pub async fn modrinth(
-    modrinth: Arc<Ferinth>,
-    config: &mut Config,
-    project_id: &str,
-) -> Result<Project> {
+pub async fn modrinth(modrinth: &Ferinth, config: &Config, project_id: &str) -> Result<Project> {
     let project = modrinth.get_project(project_id).await?;
     // Check if project has already been added
     if config.modpacks.iter().any(|modpack| {