+++ src/modpack/modrinth/structs.rs
@@ -7,17 +7,15 @@ use url::Url;
 #[serde(rename_all = "camelCase")]
 pub struct Metadata {
     /// The version of the format, stored as a number.
-    /// The current value at the time of writing is `1`
+    /// The current value at the time of writing is `1`.
     pub format_version: u64,
-    /// The game of the modpack
     pub game: Game,
-    /// A unique identifier for this specific version of the modpack
     pub version_id: String,
     /// Human readable name of the modpack
     pub name: String,
-    /// A short description of this modpack.
+    /// A short description of this modpack
     pub summary: Option<String>,
-    /// A list of files for the modpack that needs to be downloaded
+    /// A list of files for the modpack that needs
     pub files: Vec<ModpackFile>,
     /// A list of IDs and version numbers that launchers will use in order to know what to install
     pub dependencies: HashMap<DependencyID, String>,
@@ -37,9 +35,8 @@ pub enum DependencyID {
 pub struct ModpackFile {
     /// The destination path of this file, relative to the Minecraft instance directory
     pub path: PathBuf,
-    /// The hashes of the file specified
     pub hashes: ModpackFileHashes,
-    /// The  specific environment this file exists on
+    /// The specific environment this file exists on
     pub env: Option<ModpackFileEnvironment>,
     /// HTTPS URLs where this file may be downloaded
     pub downloads: Vec<Url>,