+++ src/mutex_ext.rs
@@ -1,9 +1,9 @@
 use std::sync::{Mutex, MutexGuard};
 
-/// A sketchy way to not deal with mutex poisoning
+/// An alternative lock method to ignore mutex poisoning
 ///
 /// **WARNING**: If the poison had occurred during a write, the data may be corrupted.
-/// _If_ unsafe code had poisoned the mutex, memory corruption is possible
+/// _If_ unsafe code had poisoned the mutex, memory corruption is possible.
 pub trait MutexExt<T> {
     fn force_lock(&self) -> MutexGuard<'_, T>;
 }